Created AbstractTimeBasedController class as superclass to controllers which only execute at a certain frequency
